-bugfix so that filter in all_history properly works again

+++ app/main/views.py
@@ -171,11 +171,12 @@ def all_history():
     query_has_results = True if events_query.first() else False
 
     deleteform = DeleteEventForm()
-    if deleteform.validate_on_submit():
+    if 'event_id' in request.form.values() and deleteform.validate_on_submit():
         delete_event(request.form['event_id'])
         flash('Event successfully deleted', category='success')
         current_app.logger.info('{} deleted clock event with event_id {}'
                                 .format(current_user.email, request.form['event_id']))
+        return redirect(url_for('main.clear'))
 
     current_app.logger.info('End function all_history()')
     return render_template('main/all_history.html',
